Fix undersized txpwrlmt message to firmware
The message includes not only the subband data but also the subband
header so reducing the message by the header size caused some data
read from file to trashed on way to firmware.  With this mod can read
back same subband data as is present in the txpwrlmt_cfg.conf file.