Fix *_set_key function for WPA_PSK
The fix is by incorporating Laird's code from the same functions.
Based from lrd-7.0.0.x branch of LairdCP/lrdmwl.

The problem is in WPA_PSK mode, the STA associated, but in pinging from
the STA to the AP, mwlwifi sends the ARP request ok, the AP receives it
and sends the ARP reply.  The STA does not receive the ARP reply.